fix: split reverse runner queue primitives
